Remove 'Conversion dialog'
